requirements.txt
langchain==0.2.6
langchain-huggingface==0.0.3
python-dotenv==1.0.1
sentence_transformers==3.0.1
flask==3.0.3
flask-cors==4.0.1
beautifulsoup4==4.12.3
requests==2.32.3
html2text==2024.2.26
faiss-cpu==1.8.0.post1